+++ extensions/amp-video/0.1/test/test-amp-video.js
@@ -1,4 +1,3 @@
-
 /**
  * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
  *
@@ -17,14 +16,12 @@
 
 import {createIframePromise} from '../../../../testing/iframe';
 import '../amp-video';
-import {
-  installVideoManagerForDoc,
-} from '../../../../src/service/video-manager-impl';
 import * as sinon from 'sinon';
 
-const TAG = 'amp-video2';
+const TAG = 'amp-video';
 
 describe(TAG, () => {
+
   let sandbox;
 
   beforeEach(() => {
@@ -42,7 +39,6 @@ describe(TAG, () => {
   function getVideo(attributes, children, opt_beforeLayoutCallback) {
     return createIframePromise(
         true, opt_beforeLayoutCallback).then(iframe => {
-          installVideoManagerForDoc(iframe.win.document);
           const v = iframe.doc.createElement(TAG);
           for (const key in attributes) {
             v.setAttribute(key, attributes[key]);
@@ -289,6 +285,22 @@ describe(TAG, () => {
       const impl = v.implementation_;
       expect(impl.toggleFallback.called).to.be.true;
       expect(impl.toggleFallback.calledWith(true)).to.be.true;
+    });
+  });
+
+  it('play() should not log promise rejections', () => {
+    const playPromise = Promise.reject('The play() request was interrupted');
+    const catchSpy = sandbox.spy(playPromise, 'catch');
+    return getVideo({
+      src: 'video.mp4',
+      width: 160,
+      height: 90,
+    }, null, function(element) {
+      const impl = element.implementation_;
+      sandbox.stub(impl.video_, 'play').returns(playPromise);
+      impl.play();
+    }).then(() => {
+      expect(catchSpy.called).to.be.true;
     });
   });
