+++ extensions/amp-video/0.1/validator-amp-video.protoascii
@@ -1,0 +1,85 @@
+#
+# Copyright 2016 The AMP HTML Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the license.
+#
+
+tags: {  # amp-video
+   html_format: AMP
+   html_format: AMP4ADS
+   tag_name: "SCRIPT"
+   spec_name: "amp-video extension .js script"
+   mandatory_parent: "HEAD"
+   unique: true
+   attrs: {
+     name: "async"
+     mandatory: true
+     value: ""
+   }
+   attrs: {
+     name: "custom-element"
+     mandatory: true
+     value: "amp-video"
+     dispatch_key: true
+   }
+   attrs: {
+     name: "src"
+     mandatory: true
+     value_regex: "https://cdn\\.ampproject\\.org/v0/amp-video-(latest|0\\.1)\\.js"
+   }
+   attrs: {
+     name: "type"
+     value_casei: "text/javascript"
+   }
+   cdata: {
+     blacklisted_cdata_regex: {
+       regex: "."
+       error_message: "contents"
+     }
+   }
+   spec_url: "https://www.ampproject.org/docs/reference/extended/amp-video.html"
+ }
+tags: {  # <amp-video>
+  html_format: AMP
+  html_format: AMP4ADS
+  tag_name: "AMP-VIDEO"
+  disallowed_ancestor: "AMP-SIDEBAR"
+  also_requires_tag_warning: "amp-video extension .js script"
+  attrs: { name: "alt" }
+  attrs: { name: "attribution" }
+  attrs: { name: "autoplay" value: "" }
+  attrs: { name: "controls" value: "" }
+  attrs: { name: "loop"     value: "" }
+  attrs: { name: "muted"    value: "" }
+  attrs: { name: "placeholder" }
+  attrs: { name: "poster" }
+  attrs: { name: "preload" value_regex: "(none|metadata|auto|)" }
+  attrs: {
+    name: "src"
+    value_url: {
+      allowed_protocol: "https"
+      allow_relative: true  # Will be set to false at a future date.
+    }
+    blacklisted_value_regex: "__amp_source_origin"
+  }
+  attr_lists: "extended-amp-global"
+  spec_url: "https://www.ampproject.org/docs/reference/amp-video.html"
+  amp_layout: {
+    supported_layouts: FILL
+    supported_layouts: FIXED
+    supported_layouts: FIXED_HEIGHT
+    supported_layouts: FLEX_ITEM
+    supported_layouts: NODISPLAY
+    supported_layouts: RESPONSIVE
+  }
+}
