+++ extensions/amp-video/amp-video.md
@@ -1,0 +1,106 @@
+<!---
+Copyright 2015 The AMP HTML Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS-IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+
+# <a name="amp-video"></a> `amp-video`
+
+<table>
+   <tr>
+    <td class="col-fourty"><strong>Description</strong></td>
+    <td>A replacement for the HTML5 <code>video</code> tag; only to be used for direct HTML5 video file embeds.</td>
+  </tr>
+   <tr>
+    <td class="col-fourty"><strong>Availability</strong></td>
+    <td>Stable</td>
+  </tr>
+  <!--
+  Uncomment the script after validator changes are pushed to production.
+  <tr>
+    <td width="40%"><strong>Required Script</strong></td>
+    <td><code>&lt;script async custom-element="amp-video" src="https://cdn.ampproject.org/v0/amp-video-0.1.js">&lt;/script></code></td>
+  </tr>
+  -->
+  <tr>
+    <td class="col-fourty"><strong>Examples</strong></td>
+    <td><a href="https://ampbyexample.com/components/amp-video/">Annotated code example for amp-video</a></td>
+  </tr>
+  <tr>
+    <td class="col-fourty"><strong><a href="https://www.ampproject.org/docs/guides/responsive/control_layout.html">Supported Layouts</a></strong></td>
+    <td>fill, fixed, fixed-height, flex-item, nodisplay, responsive</td>
+  </tr>
+</table>
+
+## Behavior
+
+The `amp-video` component loads the video resource specified by its `src` attribute lazily, at a time determined by the runtime. It can be controlled much the same way as a standard HTML5 `video` tag.
+
+The `amp-video` component HTML accepts up to three unique types of HTML nodes as children - `source` tags, a placeholder for before the video starts, and a fallback if the browser doesn’t support HTML5 video.
+
+`source` tag children can be used in the same way as the standard `video` tag, to specify different source files to play.
+
+One or zero immediate child nodes can have the `placeholder` attribute. If present, this node and its children form a placeholder that will display instead of the video. A click or tap anywhere inside of the `amp-video` container will replace the placeholder with the video itself.
+
+One or zero immediate child nodes can have the `fallback` attribute. If present, this node and its children form the content that will be displayed if HTML5 video is not supported on the user’s browser.
+
+## Example
+
+```html
+<amp-video width=400 height=300 src="https://yourhost.com/videos/myvideo.mp4"
+    poster="myvideo-poster.jpg">
+  <div fallback>
+    <p>Your browser doesn’t support HTML5 video</p>
+  </div>
+  <source type="video/mp4" src="foo.mp4">
+  <source type="video/webm" src="foo.webm">
+</amp-video>
+```
+
+## Attributes
+
+**src**
+
+Required if no `<source>` children are present. Must be HTTPS.
+
+**poster**
+
+The image for the frame to be displayed before video playback has started. By
+default the first frame is displayed.
+
+**autoplay**
+
+If this attribute is present, and the browser supports autoplay:
+
+* the video is automatically muted before autoplay starts
+* when the video is scrolled out of view, the video is paused
+* when the video is scrolled into view, the video resumes playback
+* when the user taps the video, the video is unmuted
+* if the user has interacted with the video (e.g., mutes/unmutes, pauses/resumes, etc.), and the video is scrolled in or out of view, the state of the video remains as how the user left it.  For example, if the user pauses the video, then scrolls the video out of view and returns to the video, the video is still paused. 
+
+**controls**
+
+Similar to the `video` tag `controls` attribute - if present, the browser offers controls to allow the user to control video playback.
+
+**loop**
+
+If present, will automatically loop the video back to the start upon reaching the end.
+
+**muted (deprecated)**
+
+`muted` attribute is deprecated and no longer has any effect.
+`autoplay` attribute automatically controls the mute behavior.
+
+## Validation
+
+See [amp-video rules](https://github.com/ampproject/amphtml/blob/master/validator/validator-main.protoascii) in the AMP validator specification.
