+++ validator/testdata/feature_tests/mandatory_dimensions.out
@@ -31,4 +31,5 @@ feature_tests/mandatory_dimensions.html:150:7 The tag 'amp-iframe extension .js 
 feature_tests/mandatory_dimensions.html:150:7 The tag 'amp-instagram extension .js script' is missing or incorrect, but required by 'amp-instagram'. (see https://www.ampproject.org/docs/reference/extended/amp-instagram.html) [AMP_TAG_PROBLEM]
 feature_tests/mandatory_dimensions.html:150:7 The tag 'amp-lightbox extension .js script' is missing or incorrect, but required by 'amp-lightbox'. (see https://www.ampproject.org/docs/reference/extended/amp-lightbox.html) [AMP_TAG_PROBLEM]
 feature_tests/mandatory_dimensions.html:150:7 The tag 'amp-twitter extension .js script' is missing or incorrect, but required by 'amp-twitter'. (see https://www.ampproject.org/docs/reference/extended/amp-twitter.html) [AMP_TAG_PROBLEM]
+feature_tests/mandatory_dimensions.html:150:7 The tag 'amp-video extension .js script' is missing or incorrect, but required by 'amp-video'. This will soon be an error. (see https://www.ampproject.org/docs/reference/amp-video.html) [DEPRECATION]
 feature_tests/mandatory_dimensions.html:150:7 The tag 'amp-youtube extension .js script' is missing or incorrect, but required by 'amp-youtube'. (see https://www.ampproject.org/docs/reference/extended/amp-youtube.html) [AMP_TAG_PROBLEM]
